Driver Corrections

+++ src/Driver/cs43l22/cs43l22.h
@@ -196,8 +196,8 @@ uint32_t cs43l22_Reset(uint16_t DeviceAddr);
 /* AUDIO IO functions */
 void      AUDIO_IO_Init(void);
 void      AUDIO_IO_DeInit(void);
-void      AUDIO_IO_Write(uint8_t Addr, uint16_t Reg, uint8_t Value);
-uint8_t   AUDIO_IO_Read(uint8_t Addr, uint16_t Reg);
+void      AUDIO_IO_Write(uint8_t Addr, uint8_t Reg, uint8_t Value);
+uint8_t   AUDIO_IO_Read(uint8_t Addr, uint8_t Reg);
 
 /* Audio driver structure */
 //extern AUDIO_DrvTypeDef   cs43l22_drv;
